requirements.txt
fastapi
psycopg2
sqlalchemy
passlib[bcrypt] 
python-jose[cryptography]
fastapi-security